Use common memref header
