+++ runtime/Makefile.rules
@@ -35,6 +35,7 @@ CFLAGS += -I$(SNITCH_RTL_PATH)/sw/math/src/internal
 CFLAGS += -I$(SNITCH_RTL_PATH)/sw/math/include/bits
 CFLAGS += -I$(SNITCH_RTL_PATH)/sw/math/include
 CFLAGS += -I$(SNITCH_RTL_PATH)/target/snitch_cluster/sw/snax/mac/include
+CFLAGS += -I$(MAKEFILE_RULES_DIRNAME)include
 CFLAGS += -D__DEFINED_uint64_t
 CFLAGS += -menable-experimental-extensions
 CFLAGS += -mcpu=snitch
